requirements.txt
django-extra-views==0.12.0
django-session-security==2.6.2
django==2.1.3
gunicorn==19.9.0
pathlib==1.0.1
psycopg2-binary==2.7.6.1
python-decouple==3.1
pytz==2018.7
six==1.11.0